+++ js/data.js
@@ -52,6 +52,7 @@ const createComment = () => ({
 });
 
 const createPictureDescription = () => {
+
   // Костыль, так как с сервера приходят данные по принципу { id: 0; url: id + 1 },
   // а здесь id и url никак не связаны были (изначально в задании 2 рандомных числа), поэтому пришлось сделать так
