+++ js/main.js
@@ -1,2 +1,2 @@
 import './render-thumb.js';
-import './full-picture.js';
+import './render-full-picture.js';
